Move allocation owner to alloc module
